Match brakeman's rules for Rails' app detection.
This may not be a preferred way, but it is how Brakeman determines
what is a valid Rails app to scan and what isn't. If we choose a
different hueristic, than you can get into a scenario where codeclimate
by default enables Brakeman (b/c a script/rails exists), and then
brakeman complains it's not a Rails app (No 'app' directory).

It's unlikely to be the case, but still worthwhile for them to match
and I see no downside.